Fixed blending bug for nested nodes:
If a subnode clears its framebuffer with a fully opaque color and then
draws a transparent image on that buffer, the default blending mode of
GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA made the framebuffer partly
transparent again. Using glBlendFuncSeparate makes it possible to split rgb
and alpha calculations. The selected parameters create the expected result:
The framebuffer stays opaque. I guess this combination should be the
default mode for most blending operations. Thanks to cupe for the help.